+++ R/model_land.R
@@ -465,7 +465,7 @@
 #'     \item{\code{channel}: A logical vector indicating whether each cell belongs to the channel network.}
 #'     \item{\code{depth_to_bedrock}: Depth to bedrock (mm).}
 #'     \item{\code{bedrock_conductivity}: Bedrock (saturated) conductivity (in m·day-1).}
-#'     \item{\code{bedrock_porosity}: Bedrock porosity ([0-1]).}
+#'     \item{\code{bedrock_porosity}: Bedrock porosity.}
 #'     \item{\code{snowpack}: A numeric vector with the snow water equivalent content of the snowpack in each cell.}
 #'     \item{\code{aquifer}: A numeric vector with the water content of the aquifer in each cell.}
 #'     \item{\code{represented_area}: Area represented by each cell (in m2).}