+++ src/aspwb.h
@@ -11,8 +11,6 @@ List aspwb_day_internal(List x, NumericVector meteovec,
               double elevation, double slope, double aspect, 
               double runon=0.0, bool verbose=false);
 
-List aspwb_day(List x, 
-               CharacterVector date, 
-               double tmin, double tmax, double rhmin, double rhmax, double rad, double wind, 
+List aspwb_day(List x, CharacterVector date, NumericVector meteovec, 
                double latitude, double elevation, double slope, double aspect,
-               double prec, double runon=0.0, bool modifyInput = true);
+               double runon=0.0, bool modifyInput = true);